% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brks_to_txt.R
\name{brks_to_txt}
\alias{brks_to_txt}
\title{Convert breaks matrix to text that can be stored in readme}
\usage{
brks_to_txt(brks)
}
\arguments{
\item{brks}{a matrix of breaks with columns start, end and class}
}
\value{
A string with breaks in the format "class: (start - end);"
}
\description{
Convert breaks matrix to text that can be stored in readme
}
\examples{
brks_to_txt(matrix(data = 1:6, nrow = 2, byrow = TRUE))
}
