% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Title}
\usage{
build_report(saved)
}
\arguments{
\item{index}{}
}
\description{
Title
}
\examples{
saved <- test_files()$saved$final2 \%>\%
  load_previous()

build_report(saved)
}
