% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{calc_ccei}
\alias{calc_ccei}
\title{Caculate CCEI from prepared rasters}
\usage{
calc_ccei(
  path_ccei = "misc/ccei",
  scenario,
  models = NULL,
  out_append = NULL,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{path_ccei}{Character. Path to pre-downloaded climate data to be used to
calculate CCEI. Expects two folders \code{historical} and \code{future}.}

\item{scenario}{Character. Which scenario to calculate CCEI for. Must match
scenario used in raster file names, e.g., "ssp245" or "ssp585"}

\item{models}{Character Vector. Subset of models to include (otherwise uses
all models).}
}
\value{
Writes final CCEI rasters (one for each scenario) to \code{path_ccei}.
\code{ccei_SCENARIO.tif}.
}
\description{
Caculate CCEI from prepared rasters
}
\examples{
calc_ccei(scenario = "ssp245")
calc_ccei(scenario = "ssp585")
calc_ccei(scenario = "ssp245",
          models = c("ACCESS-ESM1-5", "CanESM5"),
          out_append = "test")
}
