% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_raster.R
\name{calc_prop_raster}
\alias{calc_prop_raster}
\title{Get the proportion of each class in a raster that overlaps a polygon}
\usage{
calc_prop_raster(
  rast,
  poly,
  var_name,
  scope,
  val_range = 1:6,
  eer_pkg = requireNamespace("exactextractr", quietly = TRUE)
)
}
\arguments{
\item{eer_pkg}{}
}
\description{
Get the proportion of each class in a raster that overlaps a polygon
}
