% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{calc_sed}
\alias{calc_sed}
\title{Standardized Euclidean Distance for local climate}
\usage{
calc_sed(b, a, s)
}
\arguments{
\item{b}{Named List of Numeric Vectors. Vectors of values for Future climate variables.}

\item{a}{Named List of Numeric Vectors. Vectors of values for Historical climate variables.}

\item{s}{Named List of Numeric vectors. Each list item represents a vector for a
different climate variable. Values are standard deviations of the
interannual variability for historical climate variable.}
}
\description{
From Williams et al. 2007, calculate the Standardized Euclidean Distance
(SED) for local climate change for a series of raster tiles.
}
\details{
Each named vector represents a different climate variable Names must be
consistent among \code{b}, \code{a}, and \code{s}. Each value represents the value for that
climate variable for a given raster tile.
}
