% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vulnerability.R
\name{calc_vulnerability}
\alias{calc_vulnerability}
\title{Calculate the vulnerability index score}
\usage{
calc_vulnerability(df)
}
\description{
Calculate the vulnerability index score from the completed inputs. Equivalent
to the Calculations sheet in the Excel version.
}
