% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{ccei_annual}
\alias{ccei_annual}
\title{Calculate Annual Climate Moisture Deficit and Mean Temperature}
\usage{
ccei_annual(path_hist = "misc/ccei/historical", quiet = FALSE)
}
\arguments{
\item{path_hist}{Character. Path to pre-downloaded historical climate data.}
}
\value{
Annual CMD raster tif and Annual Mean Temperature raster tif saved
to an 'intermediate' folder in the \code{path_hist}.
}
\description{
Calculate Annual CMD and Mean Temperature for historical values in
preparation for creating the Climate Exposure Index Raster.
}
\details{
Annual CMD = Sum of Monthly CMD (difference between Eref and monthly
precipitation)

Annual Mean Temperature = Mean of monthly average temperature
(Midpoint: (Monthly maximum temp - Monthly minimum temp) / 2)
}
\examples{
ccei_annual()
}
