% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{ccei_monthly}
\alias{ccei_monthly}
\title{Calculate monthly Climate Moisture Deficit and Mean Temperature}
\usage{
ccei_monthly(r, month, clip, quiet)
}
\arguments{
\item{r}{SpatRaster. Historical values for one month, \code{prec}, \code{tmin}, \code{tmax}.}

\item{month}{Numeric. Which month?}

\item{clip}{SpatExtent. Area which to clip the raster to.}

\item{quiet}{Logical. Silence progress messages.}
}
\value{
SpatRaster with \code{tmean} and \code{cmd}
}
\description{
Because climr:::calc_Eref() and and climr:::calc_cmd are expected to work
on vectors, they don't work with a raster in memory.
}
\details{
terra::app() is very slow as it applies a function to every cell, where as
even if we extract the raster values to memory, being able to apply things in
parallel is much faster.

Therefore we extract raster values calculate the Eref and CMD and then return
as a raster.
}
