% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{ccei_values}
\alias{ccei_values}
\title{Calculate Annual Climate Moisture Deficit and Mean Temperature}
\usage{
ccei_values(rasts, out, aggregate = FALSE, overwrite = TRUE, quiet = FALSE)
}
\arguments{
\item{rasts}{Data.frame of rasters from \code{prep_ccei_historical()} or
\code{prep_ccei_future()}}

\item{out}{Character. Out folder/file preface}

\item{aggregate}{Logical. Whether tmean and cmd should be aggregated.}

\item{overwrite}{Logical. Whether to overwrite existing file(s)}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\value{
Annual CMD raster tif and Annual Mean Temperature raster tif saved
to an 'intermediate' folder in the \code{path_ccei}. If \code{aggregate == TRUE},
then a final raster with mean and standard deviations calculated.
}
\description{
Calculate Annual CMD and Mean Temperature in
preparation for creating the Climate Exposure Index Raster.
}
\details{
CMD = Sum of Monthly CMD (difference between Eref and monthly
precipitation)

Mean Temperature = Mean of monthly average temperature
(Midpoint: (Monthly maximum temp + Monthly minimum temp) / 2)
}
\examples{
\dontrun{
  ccei_values()
}
}
