% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{ccei_vars}
\alias{ccei_vars}
\title{Calculate monthly Climate Moisture Deficit and Mean Temperature}
\usage{
ccei_vars(prec_files, tmin_files, tmax_files, clip, quiet)
}
\arguments{
\item{prec_files}{Character. Files paths to precipitation rasters}

\item{tmin_files}{Character. Files paths to minimum temperature rasters}

\item{tmax_files}{Character. Files paths to maximum temperature rasters}

\item{clip}{SpatExtent. Area which to clip the raster to.}

\item{quiet}{Logical. Silence progress messages.}
}
\value{
SpatRaster with \code{tmean} and \code{cmd}
}
\description{
Because \code{climr:::calc_Eref()} and and \code{climr:::calc_CMD()} are expected to
work on vectors, they don't work with a raster in memory.
}
\details{
\code{terra::app()} is very slow as it applies a function to every cell, where as
even if we extract the raster values to memory, being able to apply things in
parallel is much faster.

Therefore we extract raster values calculate the Eref and CMD and then return
as a raster. It's faster if we omit the NAs for these calculations.
}
