% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_polys}
\alias{check_polys}
\title{Check and fix polygons}
\usage{
check_polys(poly, var_name = "polygon", quiet = FALSE)
}
\arguments{
\item{poly}{Spaitial object. sf or convertible to sf.}

\item{var_name}{Character. Name of the variable for messaging}
}
\value{
sf polygon
}
\description{
Check and fix polygons
}
