% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comb_index_tbl}
\alias{comb_index_tbl}
\title{Combined index lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 3 columns.
}
\usage{
comb_index_tbl
}
\description{
A table for combining index scores from the sensitivity (B/C) and modelled response
to climate change (D) sections
}
\keyword{datasets}
