% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Documentation common to different functions}
\arguments{
\item{path_hist}{Character. Path to pre-downloaded historical climate data.}

\item{path_future}{Character. Path to pre-downloaded future climate data.}

\item{workers}{}
}
\description{
Documentation common to different functions
}
\details{
Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
