% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{context_North_Am}
\alias{context_North_Am}
\title{Contextual map data for North America}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 64 rows and 2 columns.
}
\usage{
context_North_Am
}
\description{
Data used to build context map in pdf report.
}
\keyword{datasets}
