% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clim_data.R
\name{crs_albers_na}
\alias{crs_albers_na}
\title{Create the North America Albers Equal Area projection}
\usage{
crs_albers_na()
}
\value{
CRS definition
}
\description{
For area calculations it's best to use a projection which preserves area.
The Equal Areas Conic Projection is good for this, and can be customized to
North America
}
\details{
This definition is the OGC WKT from https://epsg.io/102008.
}
\examples{
crs_albers_na()
}
