% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vuln_df.R
\name{make_vuln_df}
\alias{make_vuln_df}
\title{Make vulnerability table}
\usage{
make_vuln_df(
  sp_nm,
  val1 = -1,
  val2 = NA,
  val3 = NA,
  val4 = NA,
  cave = 0,
  mig = 0,
  use_spatial = TRUE
)
}
\arguments{
\item{sp_nm}{Species name}

\item{val1}{A single number to fill the first column with. The default -1 for
Unknown should be used in most cases.}

\item{val2, val3, val4}{Additional values. Use default NA}

\item{cave}{0 or 1 For whether the species is cave or ground water dependent.
See Guidelines.}

\item{mig}{0 or 1 is the species migratory?}

\item{use_spatial}{if TRUE then values for factors that are calculated in
\code{analyze_spatial} will be set to -1 for Unknown so that they do not
override the results of the spatial analysis}
}
\value{
a data.frame that can be edited and used as input for \code{vuln_df}
in \code{\link{calc_vulnerability}}
}
\description{
Make an empty vulnerability factor table that can be filled in with the
appropriate value for each vulnerability factor. The Code corresponds to the
questions in the NatureServe Guidelines and Question is an abbreviated
version of the question. See the NatureServe Guidelines for detailed
instructions on how to score species. Values can be -1: Unknown, 0: Neutral,
1: Somewhat Increase, 2: Increase, 3: Greatly Increase.
}
\examples{
make_vuln_df("sfa", cave = 1, mig = 0)
}
