% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_evidence.R
\name{plot_evidence}
\alias{plot_evidence}
\title{Plot evidence by type}
\usage{
plot_evidence(score_df, base_size = 14)
}
\value{
ggplot2 figure
}
\description{
Plot evidence by type
}
\examples{
e <- data.frame(
  score = c(TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, TRUE),
  Code = c("B1", "B2a", "B2b", "B3", "C1", "C2ai", "C2aii"),
  evidence = c("", "Other", "Expert Opinion",
               "Expert Opinion, Literature", "",
               "Spatial Analysis", "Spatial Analysis - ccviR"))
plot_evidence(e)
plot_evidence(e[-5,])
}
