% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_index.R
\name{plot_score_index}
\alias{plot_score_index}
\title{make a visualization of the D score, B/C score and index}
\usage{
plot_score_index(b_c_score, d_score)
}
\description{
make a visualization of the D score, B/C score and index
}
