% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{prep_ccei}
\alias{prep_ccei}
\title{Prepare Climate Change Exposure Index raster}
\usage{
prep_ccei(path_ccei = "misc/ccei", overwrite = TRUE, quiet = FALSE)
}
\arguments{
\item{path_ccei}{Character. Path to pre-downloaded climate data to be used to
calculate CCEI. Expects two folders \code{historical} and \code{future}.}
}
\description{
Using the Standardized Euclidean Distance method from Williams et al., 2007
and NatureServe v3.02.1 2016.
}
\examples{
prep_ccei()

}
\references{
J.W. Williams, S.T. Jackson, J.E. Kutzbach, Projected distributions of
novel and disappearing climates by 2100 AD, Proc. Natl. Acad. Sci. U.S.A.
104 (14) 5738-5742, https://doi.org/10.1073/pnas.0606292104 (2007).

B.E. Young, E. Byers, G. Hammerson, A. Frances, L. Oliver, A. Treher,
Guidelines for Using the NatureServe Climate Change Vulnerability Index.
Release 3.02. https://www.natureserve.org/sites/default/files/guidelines_natureserveclimatechangevulnerabilityindex_r3.02_1_jun_2016.pdf (June 1st 2016)
}
