% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{prep_ccei_future}
\alias{prep_ccei_future}
\title{Title}
\usage{
prep_ccei_future(path_ccei = "misc/ccei", overwrite = TRUE, quiet = FALSE)
}
\arguments{
\item{path_ccei}{}
}
\description{
WorldClim Biclimatic variables: https://www.worldclim.org/data/bioclim.html
}
\details{
\itemize{
\item BIO1 - Annual Mean Temperature
\item 
}
}
\examples{
prep_ccei_future()
}
