% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{prep_ccei_historical}
\alias{prep_ccei_historical}
\title{Prepare Historical Data for Climate Change Exposure Index}
\usage{
prep_ccei_historical(path_ccei = "misc/ccei", overwrite = TRUE, quiet = FALSE)
}
\arguments{
\item{path_ccei}{Character. Path to pre-downloaded climate data to be used to
calculate CCEI. Expects two folders \code{historical} and \code{future}.}
}
\description{
Calculate CMD and Tmean from monthly data, calculate annual values (see
\code{ccei_annual()}). Combine and calculate mean and interannual standard
deviations for each raster cell over the entire historical record.
}
\examples{
prep_ccei_historical()
}
