% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clim_data.R
\name{prep_clim_data}
\alias{prep_clim_data}
\title{Prepare climate data}
\usage{
prep_clim_data(
  mat_norm,
  mat_fut,
  cmd_norm,
  cmd_fut,
  ccei = NULL,
  map = NULL,
  mwmt = NULL,
  mcmt = NULL,
  clim_poly = NULL,
  in_folder = NULL,
  out_folder,
  reproject = FALSE,
  overwrite = FALSE,
  scenario_name = "",
  brks = NULL,
  brks_mat = NULL,
  brks_cmd = NULL,
  brks_ccei = NULL,
  quiet = FALSE,
  n = c(1, 1)
)
}
\arguments{
\item{mat_norm, mat_fut, cmd_norm, cmd_fut, ccei, map, mwmt, mcmt, clim_poly}{file paths to find data if in_folder is not given}

\item{in_folder}{file path where files are stored. Files must be named
according to the convention described in details}

\item{out_folder}{Character. Path where the processed files will be saved.}

\item{reproject}{should the data be re-projected to lat/long? Not recommended.}

\item{overwrite}{Logical. Whether to overwrite existing file(s)}

\item{scenario_name}{Character. A string identifying the climate change
scenario(s) that will be used as a suffix for the output files.}

\item{brks}{List. Optionally supply \code{brks_mat}, \code{brks_cmd} and \code{brks_ccei}
as a list.}

\item{brks_mat, brks_cmd, brks_ccei}{a matrix containing breaks to use for
classifying mat, cmd and ccei into 6, 6 and 4 classes, respectively. See
\code{\link[raster]{reclassify}} for details on the matrix format. If NULL,
the default, the breaks will be determined using the median and half the
interquartile range}

\item{quiet}{Logical. Whether to suppress progress messages.}

\item{n}{Numeric vector. Only for use by \code{prep_clim_data_multi()} for progresses
messages and to skip unnecessary steps. Do not change.}
}
\value{
Returns a list of matrices with the breaks used to classify mat, cmd
and ccei. This list can be supplied to a future call to
\code{prep_clim_data} in order to use the same breaks for multiple climate
data sets. Processed data is saved in \code{out_folder}
}
\description{
Prepare data from raw to form needed for calculating the index. See the
NatureServe Guidelines for details on how the data is prepared.
}
\details{
Definition of input data sets and file names required in in_folder:
\describe{
\item{mat_norm:}{"MAT" mean annual temperature for the historical normal period}
\item{mat_fut:}{"MAT_2050" mean annual temperature for the future under
climate change it can be any number eg 2050, 2100}
\item{cmd_norm:}{"CMD" climate moisture deficit for the historical normal period}
\item{cmd_fut:}{"CMD_2050" climate moisture deficit for the future under
climate change it can be any number eg 2050, 2100}
\item{ccei:}{"CCEI" Climate Change Exposure Index from NatureServe website}
\item{map:}{"MAP" mean annual precipitation for the historical normal period}
\item{mwmt:}{"MWMT" mean warmest month temperature for the historical
normal period}
\item{mcmt:}{"MCMT" mean coldest month temperature for the historical
normal period}
\item{clim_poly:}{An optional shapefile with a polygon of the range of the
climate data. It will be created from the climate data if it is missing
but it is faster to provide it.}
}

Accepted raster file types are ".asc", ".tif", ".nc", ".grd", ".img", ".bil"
and ".gpkg" (see \code{spatial_file_raster()}).
}
\examples{
in_folder <- system.file("extdata/clim_files/raw", package = "ccviR")
pth_out <- fs::dir_create("processed_temp")

# use first scenario to set breaks
brks_out <- prep_clim_data(mat_norm = file.path(in_folder, "NB_norm_MAT.tif"),
                           mat_fut = file.path(in_folder, "NB_RCP.4.5_MAT.tif"),
                           cmd_norm = file.path(in_folder, "NB_norm_CMD.tif"),
                           cmd_fut = file.path(in_folder, "NB_RCP.4.5_CMD.tif"),
                           map = file.path(in_folder, "NB_norm_MAP.tif"),
                           mwmt = file.path(in_folder, "NB_norm_MWMT.tif"),
                           mcmt = file.path(in_folder, "NB_norm_MCMT.tif"),
                           out_folder = pth_out,
                           clim_poly = file.path(system.file("extdata", package = "ccviR"),
                                                 "assess_poly.shp"),
                           overwrite = TRUE,
                           scenario_name = "RCP 4.5")

prep_clim_data(mat_norm = file.path(in_folder, "NB_norm_MAT.tif"),
               mat_fut = file.path(in_folder, "NB_RCP.8.5_MAT.tif"),
               cmd_norm = file.path(in_folder, "NB_norm_CMD.tif"),
               cmd_fut = file.path(in_folder, "NB_RCP.8.5_CMD.tif"),
               out_folder = pth_out,
               clim_poly = file.path(system.file("extdata", package = "ccviR"),
                                     "assess_poly.shp"),
               overwrite = TRUE,
               scenario_name = "RCP 8.5",
               brks_mat = brks_out$brks_mat, brks_cmd = brks_out$brks_cmd,
               brks_ccei = brks_out$brks_ccei)

fs::dir_delete(pth_out)
}
\seealso{
\code{\link{get_clim_vars}} for loading the processed data.
}
