% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clim_data.R
\name{prep_clim_readme}
\alias{prep_clim_readme}
\title{Prepare README for climate data}
\usage{
prep_clim_readme(
  scenario_name,
  gcm_ensemble,
  hist_period,
  fut_period,
  emissions_scenario,
  url,
  out_folder,
  brks = NULL,
  brks_mat = NULL,
  brks_cmd = NULL,
  brks_ccei = NULL
)
}
\arguments{
\item{scenario_name}{Character. A string identifying the climate change
scenario(s) that will be used as a suffix for the output files.}

\item{gcm_ensemble}{Character vector. GCM ensemble names}

\item{hist_period}{Character vector. Historical periods used.}

\item{fut_period}{Character vector. Future periods used.}

\item{emissions_scenario}{Character vector. Emissions scenarios.}

\item{url}{Character vector. Source link}

\item{out_folder}{Character. Path where the processed files will be saved.}

\item{brks}{List. Optionally supply \code{brks_mat}, \code{brks_cmd} and \code{brks_ccei}
as a list.}

\item{brks_mat, brks_cmd, brks_ccei}{a matrix containing breaks to use for
classifying mat, cmd and ccei into 6, 6 and 4 classes, respectively. See
\code{\link[raster]{reclassify}} for details on the matrix format. If NULL,
the default, the breaks will be determined using the median and half the
interquartile range}
}
\value{
Nothing, but writes the readme to the \code{out_folder}
}
\description{
Prepare README for climate data
}
\examples{
in_folder <- system.file("extdata/clim_files/raw", package = "ccviR")
pth_out <- system.file("extdata/clim_files/processed", package = "ccviR")

# Process both scenarios at once with the same set of breaks
brks <- prep_clim_data_multi(
  mat_norm = file.path(in_folder, "NB_norm_MAT.tif"),
  mat_fut = file.path(in_folder, "NB_RCP.4.5_MAT.tif"),
  cmd_norm = file.path(in_folder, "NB_norm_CMD.tif"),
  cmd_fut = file.path(in_folder, "NB_RCP.4.5_CMD.tif"),
  map = file.path(in_folder, "NB_norm_MAP.tif"),
  mwmt = file.path(in_folder, "NB_norm_MWMT.tif"),
  mcmt = file.path(in_folder, "NB_norm_MCMT.tif"),
  out_folder = pth_out,
  clim_poly = file.path(system.file("extdata", package = "ccviR"),
                        "assess_poly.shp"),
  overwrite = TRUE,
  scenario_name = c("RCP 4.5", "RCP 8.5"))

# Add README
prep_clim_readme(
  scenario_name = c("RCP 4.5", "RCP 8.5"),
  gcm_ensemble = "AdaptWest 15 CMIP5 AOGCM Ensemble",
  hist_period = "1961-1990",
  fut_period = "2050s",
  emissions_scenario = c("RCP 4.5", "RCP 8.5"),
  url = "https://adaptwest.databasin.org/pages/adaptwest-climatena-cmip5/",
  out_dir = clim_prep,
  brks = brks)
}
