% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_data_prep.R
\name{prep_exp}
\alias{prep_exp}
\title{Prepare exposure data}
\usage{
prep_exp(rast_norm, rast_fut, file_nm, reproject = TRUE, overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
Prepare exposure data into classes based on delta and mean and
1/2 standard deviation
}
