% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_spatial.R
\name{rng_chg_mat}
\alias{rng_chg_mat}
\title{Create a range change matrix}
\usage{
rng_chg_mat(lost, maintained, gained, not_suitable = NA)
}
\arguments{
\item{lost}{Numeric vector. Two values representing the lowest and highest
values indicating range lost.}

\item{maintained}{Numeric vector. Two values representing the lowest and
highest values indicating range maintained.}

\item{gained}{Numeric vector. Two values representing the lowest and highest
values indicating range gained.}

\item{not_suitable}{Numeric vector. Two values representing the lowest and
highest values indicating range not suitable.}
}
\value{
A matrix of scores for use in \code{analyze_spatial()}
}
\description{
Creates a matrix defining the range change scores for a range change raster.
This matrix is required for the \code{hs_rcl} argument of \code{analyze_spatial()}. The
matrix indicates how the range change matrix is scored; which cell values
mean the range was lost, maintained, or gained. This is then used by
\code{analyze_spatial()} to re-calibrate the range change raster to values of 0-3
indicating NA, lost, maintained, or gained.
}
\examples{
rng_chg_mat(lost = -1,
            maintained = 0,
            gained = 1,
            not_suitable = NA)

rng_chg_mat(lost = 1,
            maintained = c(2, 6),
            gained = 7,
            not_suitable = NA)
}
