% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccvi_app.R
\name{run_ccvi_app}
\alias{run_ccvi_app}
\title{Launch the CCVI app}
\usage{
run_ccvi_app(
  file_dir = getwd(),
  launch.browser = TRUE,
  port = getOption("shiny.port"),
  test.mode = FALSE
)
}
\arguments{
\item{file_dir}{The directory to locate files from}

\item{launch.browser}{logical. Run CCVI app in browser?}

\item{port}{If launch.browser is FALSE, specify port to run CCVI app}

\item{test.mode}{Should the app be launched using shiny test.mode. Only set
to TRUE for debugging.}
}
\value{
A shiny app.
}
\description{
Launch the ccviR app to calculate the NatureServe Climate Change
Vulnerability Index. See \code{vignette("app_vignette", package = "ccviR")}
for details on how to use the app.
}
\examples{
if(interactive()){
 run_ccvi_app("demo")
}
}
