% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CCVI_calcs_SO.R
\name{run_CCVI_calcs_SO}
\alias{run_CCVI_calcs_SO}
\title{Title}
\usage{
run_CCVI_calcs_SO(
  species_nm,
  scale_nm,
  root_pth,
  force_crs = TRUE,
  eer_pkg = requireNamespace("exactextractr", quietly = TRUE)
)
}
\arguments{
\item{eer_pkg}{}
}
\value{

}
\description{
Title
}
\examples{

}
