% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_data_prep.R
\name{run_data_prep}
\alias{run_data_prep}
\title{Launch the data preparation app}
\usage{
run_data_prep(
  file_dir = getwd(),
  launch.browser = TRUE,
  port = getOption("shiny.port")
)
}
\arguments{
\item{file_dir}{The directory to locate files from or "demo" to use the demo
data included in the package.}

\item{launch.browser}{logical. Run app in browser?}

\item{port}{If launch.browser is FALSE, specify port to run CCVI app.}
}
\value{
A shiny app.
}
\description{
Launch the data preparation app for the ccviR package. See
\code{vignette("data_prep_vignette", package = "ccviR")} for details on how to use
the app.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 run_data_prep("demo")
\dontshow{\}) # examplesIf}
}
