% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_prep2.R
\name{run_data_prep2}
\alias{run_data_prep2}
\title{Launch the data preparation app}
\usage{
run_data_prep2(
  file_dir = getwd(),
  launch.browser = TRUE,
  port = getOption("shiny.port")
)
}
\arguments{
\item{file_dir}{The directory to locate files from or "demo" to use the demo
data included in the package.}

\item{launch.browser}{logical. Run app in browser?}

\item{port}{If launch.browser is FALSE, specify port to run CCVI app.}

\item{test.mode}{Should the app be launched using shiny test.mode. Only set
to TRUE for debugging.}
}
\value{
A shiny app.
}
\description{
Launch the data preparation app for the ccviR package. See
\code{vignette("data_prep_vignette", package = "ccviR")} for details on how to use
the app.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 run_data_prep2("demo")
\dontshow{\}) # examplesIf}
}
