% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_data_prep.R
\name{run_prep_data}
\alias{run_prep_data}
\title{prepare data from raw to form needed for app}
\usage{
run_prep_data(
  mat_norm,
  mat_fut,
  cmd_norm,
  cmd_fut,
  ccei,
  map,
  mwmt,
  mcmt,
  in_folder,
  out_folder,
  reproject = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
prepare data from raw to form needed for app
}
