% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial.R
\name{run_spatial}
\alias{run_spatial}
\title{Title}
\usage{
run_spatial(
  range_poly,
  scale_poly,
  clim_vars_lst,
  non_breed_poly = NULL,
  hs_rast = NULL,
  eer_pkg = requireNamespace("exactextractr", quietly = TRUE)
)
}
\arguments{
\item{eer_pkg}{}
}
\value{

}
\description{
Title
}
