% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_file_poly}
\alias{spatial_file_poly}
\title{Lists of spatial file types for polygons}
\format{
An object of class \code{character} of length 3.
}
\usage{
spatial_file_poly
}
\description{
Lists of spatial file types for polygons
}
\keyword{datasets}
