% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_file_raster}
\alias{spatial_file_raster}
\title{Lists of spatial file types for rasters}
\format{
An object of class \code{character} of length 8.
}
\usage{
spatial_file_raster
}
\description{
Lists of spatial file types for rasters
}
\keyword{datasets}
