% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_file_types}
\alias{spatial_file_types}
\title{Lists of spatial file types for all spatial files}
\format{
An object of class \code{character} of length 9.
}
\usage{
spatial_file_types
}
\description{
Lists of spatial file types for all spatial files
}
\keyword{datasets}
