% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial.R
\name{trim_ras}
\alias{trim_ras}
\title{Trim NAs from raster copied from raster package internal .memtrimlayer in
raster::trim. This is not memory safe but raster::trim takes over an hour
while this takes ~30s}
\usage{
trim_ras(r, padding = 0, values = NA, filename = "", ...)
}
\arguments{
\item{...}{}

\item{x}{}
}
\description{
Trim NAs from raster copied from raster package internal .memtrimlayer in
raster::trim. This is not memory safe but raster::trim takes over an hour
while this takes ~30s
}
