% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ui_build_table}
\alias{ui_build_table}
\title{Table to build the UI}
\format{
An object of class \code{data.frame} with 47 rows and 12 columns.
}
\usage{
ui_build_table
}
\description{
A table for identifying the update function to apply to each input. May get
more uses in the future
}
\keyword{datasets}
