% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vulnq_code_lu_tbl}
\alias{vulnq_code_lu_tbl}
\title{Lookup table for vulnerability factor codes}
\format{
An object of class \code{data.frame} with 27 rows and 4 columns.
}
\usage{
vulnq_code_lu_tbl
}
\description{
Lookup table for vulnerability factor questions, codes, maximum value, and
whether the factor is calculated by the spatial analysis.
}
\keyword{datasets}
