% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brks_to_txt.R
\name{brks_to_txt}
\alias{brks_to_txt}
\title{Convert breaks matrix to text that can be stored in readme}
\usage{
brks_to_txt(brks)
}
\description{
Convert breaks matrix to text that can be stored in readme
}
