% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gain_loss.R
\name{calc_gain_loss}
\alias{calc_gain_loss}
\title{Get the proportion of habitat gained and lost in range}
\usage{
calc_gain_loss(rast, poly)
}
\arguments{
\item{poly}{}
}
\value{

}
\description{
Get the proportion of habitat gained and lost in range
}
