% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vulnerability.R
\name{calc_vulnerability}
\alias{calc_vulnerability}
\title{Calculate the vulnerability index score}
\usage{
calc_vulnerability(exp_df, vuln_df, tax_grp, n_rnds = 1000)
}
\arguments{
\item{exp_df}{data.frame of proportion of range in different classes}

\item{vuln_df}{data.frame of answers to vulnerability questions}

\item{n_rnds}{number of monte carlo repititions for determining variability
when mutltiple values are given for a factor}
}
\description{
Calculate the vulnerability index score from the completed inputs. Equivalent
to the Calculations sheet in the Excel version.
}
