% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccviR-package.R
\docType{package}
\name{ccviR-package}
\alias{ccviR}
\alias{ccviR-package}
\title{ccviR: NatureServe CCVI in R}
\description{
Implements the NatureServe climate change vulnerability index 
  calculator in R.
}
\author{
\strong{Maintainer}: Sarah Endicott \email{sarah.endicott@canada.ca} (\href{https://orcid.org/0000-0001-9644-5343}{ORCID})

}
\keyword{internal}
