% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Documentation common to different functions}
\arguments{
\item{path_ccei}{Character. Path to pre-downloaded climate data to be used to
calculate CCEI. Expects two folders \code{historical} and \code{future}.}

\item{overwrite}{Logical. Whether to overwrite existing file(s)}

\item{quiet}{Logical. Whether to suppress progress messages.}

\item{scenario_name}{Character. A string identifying the climate change
scenario(s) that will be used as a suffix for the output files.}

\item{out_folder}{Character. Path where the processed files will be saved.}

\item{brks}{List. Optionally supply \code{brks_mat}, \code{brks_cmd} and \code{brks_ccei}
as a list.}

\item{brks_mat, brks_cmd, brks_ccei}{a matrix containing breaks to use for
classifying mat, cmd and ccei into 6, 6 and 4 classes, respectively. See
\code{\link[raster]{reclassify}} for details on the matrix format. If NULL,
the default, the breaks will be determined using the median and half the
interquartile range}
}
\description{
Documentation common to different functions
}
\details{
Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
