% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Documentation common to different functions}
\arguments{
\item{path_ccei}{Character. Path to pre-downloaded climate data to be used to
calculate CCEI. Expects two folders \code{historical} and \code{future}.}
}
\description{
Documentation common to different functions
}
\details{
Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
