% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clim_vars.R
\name{get_clim_vars}
\alias{get_clim_vars}
\title{Load climate variables}
\usage{
get_clim_vars(root_pth, scenario_names = "scn1")
}
\arguments{
\item{root_pth}{A folder location where all the climate data is stored. The
names must match one of \code{c("MAT.*tif$", "CMD.*tif$", "clim_poly.*shp",
  "MAP.*tif$", "ccei.*tif$|CCEI.*tif$","MWMT.*tif$|HTN.*tif$")} and the first
three are required.}

\item{scenario_names}{character vector with names that identify multiple
future climate scenarios. If this is supplied the raster file must include
the scenario name as a suffix to the pattern mentioned above eg. if there
are two MAT files "MAT_RCP 4.5.tif" and "MAT_RCP 8.5.tif" the scenario names
should be "RCP 4.5" and "RCP 8.5". This will happen automatically if the
scenario name is provided to \code{\link{prep_clim_data}}.}
}
\value{
A list of climate variables with names "mat", "cmd", "map", "ccei",
"htn", "clim_poly". If multiple scenarios are used mat, cmd and ccei will
be SpatRasters with one layer per scenario.
}
\description{
Load climate variables and store them in a list that can be used with
\code{\link{analyze_spatial}}. The climate variables should first be prepared
using \code{\link{prep_clim_data}}.
}
\examples{
pth <- system.file("extdata/clim_files/processed", package = "ccviR")

# scenario names
scn_nms <- c("RCP 4.5", "RCP 8.5")

get_clim_vars(pth, scn_nms)
}
