% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_q_score.R
\name{plot_q_score}
\alias{plot_q_score}
\title{Plot scores for each vulnerability question}
\usage{
plot_q_score(vuln_df)
}
\arguments{
\item{vuln_df}{the \code{vuln_df} element of the result from
\code{\link{calc_vulnerability}}.}
}
\value{
plotly graph
}
\description{
Help to visualize what factors are influencing the index. Produces a plotly
graphic where the bars are the total score for the factor after modifying it
based on exposure and the dots are the exposure multiplier. Hover to see the
name of the factor.
}
\examples{
base_pth <- system.file("extData", package = "ccviR")

clim_vars <- get_clim_vars(file.path(base_pth, "clim_files/processed"))

spat_res <- run_spatial(
  range_poly = sf::read_sf(file.path(base_pth, "rng_poly_high.shp")),
  scale_poly = sf::read_sf(file.path(base_pth, "assess_poly.shp")),
  clim_vars_lst = clim_vars,
  hs_rast = raster::raster(file.path(base_pth, "HS_rast.tif")),
  hs_rcl = matrix(c(0:7, 0, 1, 2, 2 ,2, 2, 2, 3), ncol = 2)
)

# vulnerability factor table with score 1 (somewhat increase vulnerability)
# for all factors
vuln <- make_vuln_df("test_species", val1 = 1, mig = 1)

index_vuln <- calc_vulnerability(spat_res$spat_table, vuln, "Bird")

plot_q_score(index_vuln$vuln_df)
}
