% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_index.R
\name{plot_score_index}
\alias{plot_score_index}
\title{Visualize the D score, B/C score and index}
\usage{
plot_score_index(b_c_score, d_score, n_d_factors)
}
\arguments{
\item{b_c_score}{Total score in the B and C sections.}

\item{d_score}{Total score in the D section}

\item{n_d_factors}{Number of factors scored in the D section}
}
\value{
ggplot2 graph.
}
\description{
A graph to visualize where the result falls relative to the thresholds used
to calculate index values and how the combination of the D score and B/C
score affects the index value.
}
\examples{
base_pth <- system.file("extData", package = "ccviR")

clim_vars <- get_clim_vars(file.path(base_pth, "clim_files/processed"))

spat_res <- run_spatial(
  range_poly = sf::read_sf(file.path(base_pth, "rng_poly_high.shp")),
  scale_poly = sf::read_sf(file.path(base_pth, "assess_poly.shp")),
  clim_vars_lst = clim_vars,
  hs_rast = raster::raster(file.path(base_pth, "HS_rast_high.tif")),
  hs_rcl = matrix(c(0:7, 0, 1, 2, 2 ,2, 2, 2, 3), ncol = 2)
)

# vulnerability factor table with score 1 (somewhat increase vulnerability)
# for all factors
vuln <- make_vuln_df("test_species", val1 = 1, mig = 1)

index_vuln <- calc_vulnerability(spat_res$spat_table, vuln, "Bird")

plot_score_index(index_vuln$b_c_score, index_vuln$d_score, index_vuln$n_d_factors)
}
