% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_index.R
\name{plot_score_index}
\alias{plot_score_index}
\title{Visualize the D score, B/C score and index}
\usage{
plot_score_index(score_df)
}
\arguments{
\item{score_df}{A dataframe containing the result of a call to
\code{calc_vulnerability}.}
}
\value{
ggplot2 graph.
}
\description{
A graph to visualize where the result falls relative to the thresholds used
to calculate index values and how the combination of the D score and B/C
score affects the overall index value.
}
\details{
The colours show the location of the thresholds used to determine the index,
the points show the score for each scenario and the lines show the range of
scores produced by the Monte Carlo simulations. Multiple scenarios are
identified by different symbols
}
\examples{
base_pth <- system.file("extData", package = "ccviR")

# scenario names
scn_nms <- c("RCP 4.5", "RCP 8.5")

clim_vars <- get_clim_vars(file.path(base_pth, "clim_files/processed"),
                           scenario_names = scn_nms)

spat_res <- analyze_spatial(
  range_poly = sf::read_sf(file.path(base_pth, "rng_poly.shp"), agr = "constant"),
  scale_poly = sf::read_sf(file.path(base_pth, "assess_poly.shp"), agr = "constant"),
  clim_vars_lst = clim_vars,
  hs_rast = raster::stack(raster::raster(file.path(base_pth, "rng_chg_45.tif")),
                          raster::raster(file.path(base_pth, "rng_chg_85.tif"))),
  hs_rcl = matrix(c(0:7, 0, 1, 2, 2 ,2, 2, 2, 3), ncol = 2),
  scenario_names = scn_nms
)

# vulnerability factor table with score 1 (somewhat increase vulnerability)
# for all factors
vuln <- make_vuln_df("test_species", val1 = 1, mig = 1)
vuln$Value2[c(5, 7, 9)] <- 3

index_vuln <- calc_vulnerability(spat_res$spat_table, vuln, "Bird")

plot_score_index(index_vuln)

}
