% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{prep_ccei}
\alias{prep_ccei}
\title{Prepare Climate Change Exposure Index raster}
\usage{
prep_ccei(path_hist = "misc/ccei/historical", path_future = "misc/ccei/future")
}
\arguments{
\item{path_hist}{Character. Path to pre-downloaded historical climate data.}

\item{path_future}{Character. Path to pre-downloaded future climate data.}
}
\description{
Using the Standardized Euclidean Distance method from Williams et al., 2007
}
\examples{
prep_ccei()

}
\references{
J.W. Williams, S.T. Jackson, J.E. Kutzbach, Projected distributions of
novel and disappearing climates by 2100 AD, Proc. Natl. Acad. Sci. U.S.A.
104 (14) 5738-5742, https://doi.org/10.1073/pnas.0606292104 (2007).
}
