% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ccei.R
\name{prep_ccei_historical}
\alias{prep_ccei_historical}
\title{Prepare Historical Data for Climate Change Exposure Index}
\usage{
prep_ccei_historical(path_ccei = "misc/ccei", overwrite = TRUE, quiet = FALSE)
}
\arguments{
\item{path_ccei}{Character. Path to pre-downloaded climate data to be used to
calculate CCEI. Expects two folders \code{historical} and \code{future}.}
}
\value{
Writes intermediate and final rasters to \code{path_ccei} in the
"Intermediate" folder. Annual means (\code{hist_YYYY.tif} and
\code{hist_groups_VAR.tif})as well as overall historical averages and standard
deviations \code{hist_all_vars.tif}.
}
\description{
Calculate CMD and Tmean from monthly data, calculate annual values (see
\code{ccei_annual()}). Combine and calculate mean and interannual standard
deviations for each raster cell over the entire historical record.
}
\examples{
\dontrun{
  prep_ccei_historical()
}
}
