% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial.R
\name{run_spatial}
\alias{run_spatial}
\title{Run the spatial analysis functions to get inputs to index calculator}
\usage{
run_spatial(
  range_poly,
  scale_poly,
  clim_vars_lst,
  non_breed_poly = NULL,
  hs_rast = NULL
)
}
\arguments{
\item{clim_vars_lst}{}
}
\value{

}
\description{
Run the spatial analysis functions to get inputs to index calculator
}
